[settlement creation] refactor to be possible to add institutional settlement creation
