+++ README.md
@@ -13,7 +13,7 @@ Mono repository for Validator Bonds product
   ([SDK](./packages/validator-bonds-sdk/), [CLI](./packages/validator-bonds-cli/))
 * [`api/`](./api/) - in Rust developed OpenAPI service that publishes bonds data ([API endpoint](https://validator-bonds-api.marinade.finance/docs))
 * [`.buildkite/`](./.buildkite/) - automated pipelines that prepare data for bonds claiming, updating API data and similar
-* [`settlement-engine/`](protected-event-distribution/) - code for a CLI creating protected event data that is published in form of JSON and packed as a settlement on-chain
+* [`settlement-engine/`](settlement-distributions/bid-psr-distribution/) - code for a CLI creating protected event data that is published in form of JSON and packed as a settlement on-chain
 * [`merkle-tree/`](./merkle-tree/) - generic Rust library implementing the merkle tree data structure management
 * [`migrations/`](./migrations/) - SQL scripts to prepare and change DB schemas
 * [`scripts/`](./scripts/) - scripts used in pipeline and to manage and integrate various repository parts
@@ -95,7 +95,7 @@ pnpm build
 # testing the SDK+CLI against the bankrun and local validator
 pnpm test
 # running single cargo test
-cargo test --package protected-event-distribution ts_cross_check_hash_generate
+cargo test --package bid-psr-distribution ts_cross_check_hash_generate
 # bankrun part of the tests
 pnpm test:bankrun
 # local validator part of the tests
