+++ settlement-distributions/bid-distribution/src/bin/cli.rs
@@ -1,14 +1,14 @@
 use bid_distribution::sam_meta::ValidatorSamMeta;
 use bid_distribution::settlement_claims::generate_bid_settlement_collection;
 use bid_distribution::settlement_config::SettlementConfig;
+use bid_psr_distribution::merkle_tree_collection::generate_merkle_tree_collection;
+use bid_psr_distribution::settlement_collection::SettlementFunder;
+use bid_psr_distribution::settlement_collection::SettlementMeta;
+use bid_psr_distribution::settlement_config::no_filter;
+use bid_psr_distribution::settlement_config::stake_authorities_filter;
+use bid_psr_distribution::stake_meta_index::StakeMetaIndex;
+use bid_psr_distribution::utils::{read_from_json_file, write_to_json_file};
 use env_logger::{Builder, Env};
-use protected_event_distribution::merkle_tree_collection::generate_merkle_tree_collection;
-use protected_event_distribution::settlement_claims::SettlementFunder;
-use protected_event_distribution::settlement_claims::SettlementMeta;
-use protected_event_distribution::settlement_config::no_filter;
-use protected_event_distribution::settlement_config::stake_authorities_filter;
-use protected_event_distribution::stake_meta_index::StakeMetaIndex;
-use protected_event_distribution::utils::{read_from_json_file, write_to_json_file};
 use snapshot_parser_types::stake_meta::StakeMetaCollection;
 use solana_sdk::pubkey::Pubkey;
 use std::collections::HashSet;
