+++ settlement-distributions/bid-psr-distribution/src/lib.rs
@@ -1,7 +1,8 @@
 pub mod merkle_tree_collection;
 pub mod protected_events;
 pub mod revenue_expectation_meta;
-pub mod settlement_claims;
+pub mod settlement_collection;
 pub mod settlement_config;
+pub mod settlement_generator;
 pub mod stake_meta_index;
 pub mod utils;
