+++ settlement-pipelines/Cargo.toml
@@ -49,7 +49,7 @@ regex = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
-protected-event-distribution = { workspace = true }
+bid-psr-distribution = { workspace = true }
 shellexpand = { workspace = true }
 solana-sdk = { workspace = true }
 solana-client = { workspace = true }
