+++ settlement-pipelines/src/bin/init_settlement.rs
@@ -45,11 +45,11 @@ struct Args {
     #[clap(flatten)]
     global_opts: GlobalOpts,
 
-    /// Merkle tree collection file path generated by protected-event-distribution as of --output-merkle-tree-collection
+    /// Merkle tree collection file path generated by bid-psr-distribution as of --output-merkle-tree-collection
     #[arg(short = 'm', long)]
     input_merkle_tree_collection: PathBuf,
 
-    /// Settlement collection file path generated by protected-event-distribution as of --output-settlement-collection
+    /// Settlement collection file path generated by bid-psr-distribution as of --output-settlement-collection
     #[arg(short = 's', long)]
     input_settlement_collection: PathBuf,
